Dr_chatbot/requirements.txt
# Core dependencies for Dr. Meenakshi Tomar Dental Chatbot
openai>=1.3.0
pinecone>=3.0.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0

# Enhanced Processing (OpenAI-based only)
tiktoken>=0.5.0

# Quality and Monitoring
pydantic>=2.0.0
loguru>=0.7.0
agentops>=0.2.0